refactor: Rule.interface 를 abstract class로 변경
