+++ src/main/java/chess/domain/Rule.java
@@ -1,12 +1,55 @@
 package chess.domain;
 
-public interface Rule {
+public abstract class Rule {
+    public enum Type {
+        PAWN("Pawn", 1),
+        QUEEN("Queen", 9),
+        ROOK("Rook", 5),
+        KNIGHT("Knight", 2.5),
+        KING("King", 0),
+        EMPTY("Empty", 0),
+        BISHOP("Bishop", 3);
 
-    boolean isValidMove(final Position origin, final Position target);
+        private final String name;
+        private final double score;
 
-    boolean isValidAttack(final Position origin, final Position target);
+        Type(final String name, final double score) {
+            this.name = name;
+            this.score = score;
+        }
 
-    double getScore();
+        public String getName() {
+            return name;
+        }
 
-    String getName();
+        public double getScore() {
+            return score;
+        }
+    }
+
+    private Type type;
+
+    public Rule(final Type type) {
+        this.type = type;
+    }
+
+    public boolean isValidMove(final Position origin, final Position target) {
+        return false;
+    }
+
+    public boolean isValidAttack(final Position origin, final Position target) {
+        return isValidMove(origin, target);
+    }
+
+    public double getScore() {
+        return this.type.score;
+    }
+
+    public String getName() {
+        return this.type.name;
+    }
+
+    public Type getType() {
+        return type;
+    }
 }
