+++ src/main/java/chess/domain/rule/Rook.java
@@ -1,8 +1,9 @@
 package chess.domain.rule;
 
 import chess.domain.Position;
+import chess.domain.Rule;
 
-public class Rook extends AbstractRule {
+public class Rook extends Rule {
     private static Rook INSTANCE = new Rook();
 
     private Rook() {
